enabled handling of large id_array value (>2.4B)
git-svn-id: https://svn.urbansim.org/src/trunk@18935 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/datasets/abstract_dataset.py
@@ -1744,8 +1744,8 @@ class AbstractDataset(object):
                 self.id_mapping = array([], dtype="int32")
             return
         if id_array.ndim == 1:
-            maxid = id_array.max()
-            minid = id_array.min()
+            maxid = long(id_array.max())
+            minid = long(id_array.min())
             if (maxid-minid+1) <= (2*self.size()+1000):
                 self.id_mapping_type="A" #array
                 self.id_mapping_shift = minid
